# Initialise data mask with expected number of elements