Closes #1048. Performance is improved but not dramatically.

```r
data <- as.list(set_names(1:1e5))
bench::mark(before = as_data_mask(data))[1:8]
```

Before

```r
#>   expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 before        220ms    223ms      4.47    5.44MB        0     3     0
```

After

```r
#>   expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 after         169ms    170ms      5.88    2.76MB        0     3     0
```

For 1e6 elements this goes from 2.35s to 1.85s.